add .prm.out files to make eclipse happy
Even though the .out files get generated by the Makefile in doc/ when
you build manual.pdf, it turns out eclipse is complaining when the .out
files are not present.
Also remove the .out from .gitignore again and do not generate a .out
for doc/manual/empty.prm because it is not used.