rework linear solver tolerance
Change the Stokes linear solver tolerance to inlcude the current
pressure instead of only the rhs. In the past the formula for the Stokes
solver tolerance was:
   rhs.l2_norm() * parameters.linear_stokes_solver_tolerance

This causes problems if the magnitude of the rhs is large (for example
due to a large gravity vector) and if the current pressure solves the
problem fairly well (so the initial residual is much smaller than the
tolerance as defined above). This is basically caused by the static
pressure being a good solution. The new formula is:
  || B^T p_0 - G || * parameters.linear_stokes_solver_tolerance

This fixes the problem that one had to use a linear solver tolerance of
1e-15 to make convection_box.prm work.