Internal heating statistics computes mass averages.
Internal heating statistics should compute mass averages and integrate
over mass instead of volume since the internal heating rate is also
given as specific heating rate (W/kg instead of W/m^3). In the non-
dimensional testcases this was not a problem so far, because the
density was commonly assumed to be 1 and constant.