Also allow projecting to a Q1 space.
Currently, we project from quadrature points to a Q1 space, and then evaluate
back to the quadrature points. The values at the quadrature points are then
limited to the minimal and maximal values of the original data set again. This
is necessary to ensure that the density remains positive.

The whole approach does not work, as shown by the number of iterations in
manual.tex -- it is by far the worst approach. This may be due to the fact
that the Q1 projection may have negative values, even if cutting them off
in a later stage.