make test suite opt in
With this change, only a single test is added to the list of tests that
are going to be executed when you run "ctest" or "make test". To enable
all tests, either execute "make setup_tests" or configure with -D
ASPECT_RUN_ALL_TESTS=ON.

The reason for this change is, that many tests fail due to round-off
etc.. This confuses new users who expect "make; make test" to succeed.