Fix cookbooks so they can run on odd Macs as well.
For reasons that are beyond a solid understanding, the function parser
thinks that 4 or more digits are not a number on some Mac OS X
systems. The problem is that it thinks that the locale *requires*,
not just allows, the use of commas as thousands separators. It then
thinks that such numbers are obviously names, and complains that
the symbol has not been assigned a numerical value. Very awkward to
make sense of.

This only appears to be a problem inside the function parser. All other
parameters that are not treated as function expressions are ok.