fix tests/terminate_user_request
The test terminate_user_request terminates when finding the file
restart.mesh which is generated after a couple of time steps. If this
test is run again from a directory that is not clean, it will fail if
restart.mesh still exists. Fix this deleting this file by loading a
plugin upon loading.