fix hdf5 output after resume
Bug reported by Siqi Zhang on the ASPECT mailing list. HDF5 output might
be corrupted after resuming from a snapshot when the mesh file is not
written. This is likely happening because the ordering of vertices or
some other data structure can be different when we reconstruct the mesh
after resuming.
Fix this by writing the mesh data after we resume from a snapshot.