Modularize assembly.
Introduce a structure that contains signals that point to functions
that do the actual assembly loop. The setup of scratch objects etc
continues to happen at an outer level, but on every cell we then
simply call each slot in the signal, and thereby add up terms as
appropriate for an equation or approximation.