Improve an error message.
If we can't parse a function expression provided in the input file, the
previous error message simply said that there was an error but nothing
more. What was easy to miss was that there is actually more below, since
we re-throw the original error message. The improvement made here simply
points out that there is more about what caused the error below the first
error message. This now looks like this:

ERROR: FunctionParser failed to parse
        'Compositional initial conditions.Function'
with expression
        '1'
More information about the cause of the parse error
is shown below.

----------------------------------------------------
Exception on processing:

--------------------------------------------------------
An error occurred in line <76> of file </home/bangerth/p/deal.II/1/dealii/source/base/function_parser.cc> in function
    void dealii::FunctionParser<dim>::initialize(const string&, const std::vector<std::basic_string<char> >&, const std::map<std::basic_string<char>, double>&, bool) [with int dim = 3; std::string = std::basic_string<char>]
The violated condition was:
    this->n_components == expressions.size()
The name and call sequence of the exception was:
    ExcInvalidExpressionSize(this->n_components, expressions.size())
Additional Information:
The number of components (2) is not equal to the number of expressions (1).
--------------------------------------------------------

The commit also makes a couple of places more consistent.