Introspection 2.0
This PR is reworking the Introspection class to be more flexible with
the goal to a) simplify the code and b) allow the melt transport (or
others) to plug into the creation of the introspection.

The new logic is as follows:
1. a std::vector<VariableDeclaration> is constructed that defines the set
of variables to use. This includes user-friendly names, definition of
the FE space, components, and blocks.
2. A signal is fired in construct_variables (core.cc) to allow
modification of the variables.
3. Finally, the Introspection object is created from that std::vector. A
new base class (FEVariableCollection) allows for queries of information
about variables by name. Each variable is exposed by a struct of type
FEVariable.