Avoid warning about potentially unused variable.
In optimized mode, we did not check that a point at which we want to evaluate
the solution was actually found. This led to a warning about an unused
variable 'point_found'. This could be addressed by doing a 'void' cast
of said variable, but the better solution appears to be to actually
perform the check also in optimized mode -- after all, the list of points
at which we want to evaluate the solution is read at run-time from the
input field.