Fix a directory problem.
#1074 used the old directory structure in the tests/ subdirectories,
but was pre-empted by #1086 and so tests fail because the pull request
still put output files into the old location, but the executable now
places them into a different location. Fix this by moving a file to
its now correct location. It here also requires updating the .cc file
slightly.