Added a reinit function to MaterialModelInputs.
This patch adds a function to reinitialize and populate the arrays created by
the MaterialModelInputs constructor. The MaterialModelInputs constructor with the
same arguments creates the arrays, then calls this function to populate them.
This function is intended to be called within loops over cells so that arrays
are not recreated more than necessary.