Ensure that the base class also gets to serialize its data.
The current implementation of the various serialization functions of particle output
classes have a comment that says that we ought to call the serialization function
of the base class -- but then they all do not actually do so. Fix this.

It doesn't current make any difference since the base class has nothing to serialize.
But it may acquire some data at some point in the future, and this patch prevents
the bug that would then become apparent.