Improve an error message.
In particular, let the assertion check exactly the cases we support. It seems
to me that at one point, we allowed arbitrary opening angles for the spherical
shell geometry in 2d, and the assertion suggests that too, but the current
code does not allow for that. Match the error message to the actual code.