avoid deadlock when exception is thrown
If an exception is thrown, we will call the TimerOutput destructor,
which will exit subsections and in turn call Timer functions that
trigger MPI communication. This can a) hang if only some processors
throw, or b) report incorrect messages (about Trilinos compress()).

Avoid this problem by a) reseting the timer in the destructor and b)
make sure MPI_InitFinalize will not hang by moving it into the try
block.