Move Lookup namespace from grain_size to material_model/utilities
This change moves the entire Lookup namespace, which contains the class
MaterialLookup from grain_size.cc and grain_size.h over to
material_model/utilities.cc and utilities.h. The functions contained in
this namespace are used in the grain_size and steinberger material
models, and potentially more in the future. The changes to the
steinberger material model are part of separate commit.