Provide a base class with a virtual destructor.
The class has a virtual member function but a non-virtual destructor. This is generally
a bad idea, though it doesn't matter here because we never store pointers to this
particular type of class (we use this class as one of several base classes of
the viz postprocessor classes, and only store pointers to the Interface class). But
it is bad practice, so fix this anyway.