Fix: avoid MPI deadlock in constraints
In rare cases, the code to check if the ConstraintMatrix changed can
cause a deadlock (only some processes would invoke the MPI::sum to check
if the mesh changed).
For this to happen, the total number of DoFs has to stay the same during
refinement. Fix this by unconditionally doing the communication step.