Use diagonal strengthing in advection solve
Especially with SUPG, the advection/temperature solve can fail to
converge (maybe the matrix is not diagonally dominant without entropy
viscosity). Now, if the solver fails to converge, set up the ILU
preconditioner with diagonal strengthening and try again. Don't do
this by default as it increases iteration counts in normal
situations..