disable part of mesh smoothing in refine/coarse
The mesh smoothing algorithms in deal.II change coarsening/refinement
flags but lead to a hang/crash when combined with periodic boundaries.

As a work-around, disable one particular smoothing flag
eliminate_refined_boundary_islands. This flag is not important to have
regardless. Strictly speaking, this work-around is not necessary with
9.3.0 onwards.

Finally, eliminate the logic for limit_level_difference_at_vertices. It
turns out that this is contained in smoothing_on_refinement already.

fixes #3604