AsciiDataLookup: allow directly initializing
Allow specifying the data in the AsciiDataLookup directly via a call to
reinit() instead of going through parsing a .txt file. This will allow
us to offer different ways to load data (binary, etc.) and doesn't
require the opendap reader to generate a text document.

- extract the initialization from load_file into new reinit()
- let load_file use the new function
- add a unit test
- add a second argument to the private compute_table_indices function as
the dimensions are not set when it is used in load_file.

I am not too happy with the fact that we have to specify a dim Table for
each coordinate value, but this is what it is right now.