add failing Newton Stokes test
This currently failing test is a linear problem, so the second nonlinear
iteration should do 0 iterations (and does when you use "single
Advection, iterated Stokes") but it crashes with the Stokes solver not
converging. It looks like we try to reduce the residual to 1e-20, which
won't work.
The test passes if you remove "set Linear solver tolerance".