Also allow setting PERPLEX_DIR on the command line.
In what is probably a historical accident, our CMakeLists.txt file allows querying if
a user has set PERPLEX_DIR as an environment variable, but not on the command line
via
  -DPERPLEX_DIR=/where/I/installed/perplex
I suspect that that was not on purpose, and it is easily fixed. It also makes this
call to FIND_PACKAGE look more like the other calls to this command in this file.