fix benchmarks/newton/spiegelman
fixes
An error occurred in line <174> of file
<../source/boundary_temperature/interface.cc> in function
void
aspect::BoundaryTemperature::Manager<dim>::parse_parameters(dealii::ParameterHandler&)
[with int dim = 2]
The violated condition was:
    model_names.size() == 0
Additional information:
You have indicated that you wish to apply a boundary temperature
model, but the <Fixed temperature boundary indicators> parameter is
empty. Please use this parameter to specify the boundaries on which the
model(s) should be applied.