Make code in heat_flux_map.cc more efficient.
The existing code allocates a sizable array for every cell, locally owned or not,
on the boundary or not. But we end up only using the array for cells that are
locally owned and on the boundary. As a consequence, we can move the allocation
into the innermost if() condition that matters. In the process, we can also
get rid of another use of GeometryInfo.