QLS particle interpolation limiting scheme
The limiting scheme follows a very similar idea to the one already in
bilinear least squares with the addition of extension and boundary
extrapolation. The QLS limiting scheme also considers particles of
neighboring cells when deterimining the property min and max. The
limiting scheme can extrapolate an approximate value for the boundary of
a model from the cell averages of the cell that is touching the boundary
and the cell on the opposing side.