Elasticity: unconditionally fill elastic outputs
We used to only fill elastic outputs when reaction terms are requested
(meaning strain_rate is provided). This seems to be a bug, because we
don't need the strain rate for this.

This probably fixes a bug in elasticity with GMG, where we do not
request the strain rate in all cases.