GMG: shared MatrixFree objects
Share the MatrixFree objects between StokesOperator, ABlockOperator, and
SchurcomplementOperator for GMG.

This forces batching and iteration of cells in the same order and fixes
a problem that in certain situations wrong CellData (viscosity etc.) was
being read leading to slow solver convergence (or no convergence).

This also reduces setup cost and memory consumption and requires less
code. Win, win, win!?