change usage of MPI::broadcast
The current version produces warnings of the kind
```
/ssd/aspect-git2/source/simulator/parameters.cc:1629:53: warning:
‘error_code’ may be used uninitialized [-Wmaybe-uninitialized]
```
on my machine. There is no reason to use the broadcast() function from
deal.II that is made for sending complicated objects that need to be
packed if we just want to transfer an int. We could use MPI_Bcast() but
I don't like manual error checking, which this new overload does
(otherwise it is identical to MPI_Bcast here).