GMG: support no averaging for GMG
Also allow GMG to not require viscosity averaging. This has the same
memory requirement and computational cost per iteration as Q1 averaging.
Note that we still need to average the viscosity for the GMG hierarchy,
which likely results in higher iteration counts.
Case in point for nsinker (coarsest, 148k DoFs):
- active=Q0, GMG=Q0: 23 iterations
- active=none, GMG=Q0: 42 iterations
- active=none, GMG=Q1: 124 iterations